Class Address and Class Collaborator.
Fixed atribute validations according to its error.

Each variable has now its validation method which is made on the "set" method and constructor now has the sets on it.

Added enum type for the diferent results that may happen on Postal Code and Collaborator name (so the error message can be specified for a specific error instead of a "generic error message".