US004: SD and Rationale
Added CollaboratorRepository to the System Sequence as we need to Update the selected collaborator on the list of Collaborators.
Also added the methods that will be used in CollaboratorRepository to validate if the selected collaborator already has that skill assigned and the method to actually add skills to collaborator.

Updated Rationale Table and Systematization accordingly.