US005 Class TeamProposalUI:
Implemented the first steps in this class.
Implemented all the methods to read the inputs, including the Set Of Skills (Skill list and quantity). Methods have try catch already, to prevent some input mismatch.

Still need to implement the submit data / generate a team to be accepted or rejected.

US005 System Sequence:
Added a new file (SD-testing) because i'll try to make new updates in the future without changing the original SD.